refactor: move slot into it's own object; capitalize class names
