+++ objects/osuToken.py
@@ -27,57 +27,6 @@ if TYPE_CHECKING:
 
 
 class token:
-    __slots__ = (
-        "userID",
-        "username",
-        "safeUsername",
-        "privileges",
-        "whitelist",
-        "staff",
-        "restricted",
-        "irc",
-        "kicked",
-        "loginTime",
-        "pingTime",
-        "timeOffset",
-        "streams",
-        "tournament",
-        "messagesBuffer",
-        "blockNonFriendsDM",
-        "spectators",
-        "spectating",
-        "spectatingUserID",
-        "location",
-        "joinedChannels",
-        "ip",
-        "country",
-        "awayMessage",
-        "sentAway",
-        "matchID",
-        "tillerino",
-        "silenceEndTime",
-        "protocolVersion",
-        "queue",
-        "spamRate",
-        "actionID",
-        "actionText",
-        "actionMd5",
-        "actionMods",
-        "gameMode",
-        "relax",
-        "autopilot",
-        "beatmapID",
-        "rankedScore",
-        "accuracy",
-        "playcount",
-        "totalScore",
-        "gameRank",
-        "pp",
-        "token",
-        "processingLock",
-        "_bufferLock",
-        "_spectLock",
-    )
 
     def __init__(
         self,
